Refactors GithubNavigation.kt
